requirements.txt
numpy
scipy
scikit-learn
networkx
umap-learn
squidpy
statsmodels